requirements.txt
discord.py>=2.3.0
aiosqlite>=0.19.0
chat-exporter>=2.8.0
